refactor: move incrementalRestoreActionBlacklist config
