--- br/pkg/restore/config.go
@@ -1,10 +1,0 @@
-package restore
-
-import "github.com/pingcap/tidb/parser/model"
-
-var incrementalRestoreActionBlacklist = []model.ActionType{
-	model.ActionSetTiFlashReplica,
-	model.ActionUpdateTiFlashReplicaStatus,
-	model.ActionLockTable,
-	model.ActionUnlockTable,
-}
